# swagger/spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restaurant Review API",
    "version": "1.0.0",
    "description": "A comprehensive restaurant review and rating system API",
    "contact": {
      "name": "API Support",
      "email": "support@restaurantreview.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/restaurants": {
      "get": {
        "tags": ["Restaurants"],
        "summary": "Get all restaurants",
        "description": "Retrieve a list of all restaurants with optional filtering",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search term for restaurant name or description",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "Filter by city",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "Filter by category ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "price_range",
            "in": "query",
            "description": "Filter by price range (1-4)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 4
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of restaurants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Restaurant"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Restaurants"],
        "summary": "Create a new restaurant",
        "description": "Add a new restaurant to the system",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRestaurantRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Restaurant created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurant"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/restaurants/{id}": {
      "get": {
        "tags": ["Restaurants"],
        "summary": "Get restaurant by ID",
        "description": "Retrieve a specific restaurant with its reviews",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Restaurant ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restaurant details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurant"
                }
              }
            }
          },
          "404": {
            "description": "Restaurant not found"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories",
        "description": "Retrieve a list of all restaurant categories",
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Get all reviews",
        "description": "Retrieve a list of reviews with optional filtering",
        "parameters": [
          {
            "name": "restaurant_id",
            "in": "query",
            "description": "Filter by restaurant ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "integer",
            "description": "Unique identifier for the category"
          },
          "category_name": {
            "type": "string",
            "description": "Name of the category"
          },
          "category_description": {
            "type": "string",
            "nullable": true,
            "description": "Description of the category"
          },
          "is_active": {
            "type": "boolean",
            "default": true,
            "description": "Whether the category is active"
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "description": "When the category was created"
          }
        }
      },
      "Restaurant": {
        "type": "object",
        "properties": {
          "restaurant_id": {
            "type": "integer",
            "description": "Unique identifier for the restaurant"
          },
          "name": {
            "type": "string",
            "description": "Restaurant name"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Restaurant description"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "description": "Street address"
          },
          "city": {
            "type": "string",
            "nullable": true,
            "description": "City"
          },
          "state": {
            "type": "string",
            "nullable": true,
            "description": "State or province"
          },
          "zip_code": {
            "type": "string",
            "nullable": true,
            "description": "ZIP or postal code"
          },
          "website": {
            "type": "string",
            "nullable": true,
            "description": "Restaurant website URL"
          },
          "hours_operation": {
            "type": "object",
            "nullable": true,
            "description": "Operating hours for each day"
          },
          "price_range": {
            "type": "integer",
            "nullable": true,
            "minimum": 1,
            "maximum": 4,
            "description": "Price range (1=$ to 4=$$$$)"
          },
          "average_rating": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 5,
            "description": "Average rating out of 5"
          },
          "total_reviews": {
            "type": "integer",
            "nullable": true,
            "description": "Total number of reviews"
          },
          "latitude": {
            "type": "number",
            "nullable": true,
            "description": "Latitude coordinate"
          },
          "longitude": {
            "type": "number",
            "nullable": true,
            "description": "Longitude coordinate"
          },
          "is_verified": {
            "type": "boolean",
            "default": false,
            "description": "Whether the restaurant is verified"
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "description": "When the restaurant was created"
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the restaurant was last updated"
          },
          "is_active": {
            "type": "boolean",
            "default": true,
            "description": "Whether the restaurant is active"
          }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "review_id": {
            "type": "integer",
            "description": "Unique identifier for the review"
          },
          "user_id": {
            "type": "integer",
            "description": "ID of the user who wrote the review"
          },
          "restaurant_id": {
            "type": "integer",
            "description": "ID of the restaurant being reviewed"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "Rating out of 5"
          },
          "review_title": {
            "type": "string",
            "nullable": true,
            "description": "Title of the review"
          },
          "review_text": {
            "type": "string",
            "nullable": true,
            "description": "Review content"
          },
          "is_recommended": {
            "type": "boolean",
            "nullable": true,
            "description": "Whether the user recommends the restaurant"
          },
          "moderation_status": {
            "type": "string",
            "nullable": true,
            "enum": ["pending", "approved", "rejected"],
            "description": "Moderation status of the review"
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "description": "When the review was created"
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the review was last updated"
          },
          "is_active": {
            "type": "boolean",
            "default": true,
            "description": "Whether the review is active"
          }
        }
      },
      "CreateRestaurantRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Restaurant name"
          },
          "description": {
            "type": "string",
            "description": "Restaurant description"
          },
          "address": {
            "type": "string",
            "description": "Street address"
          },
          "city": {
            "type": "string",
            "description": "City"
          },
          "state": {
            "type": "string",
            "description": "State or province"
          },
          "zip_code": {
            "type": "string",
            "description": "ZIP or postal code"
          },
          "website": {
            "type": "string",
            "description": "Restaurant website URL"
          },
          "hours_operation": {
            "type": "object",
            "description": "Operating hours for each day"
          },
          "price_range": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4,
            "description": "Price range (1=$ to 4=$$$$)"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude coordinate"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude coordinate"
          },
          "category_ids": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "Array of category IDs"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          }
        }
      }
    }
  }
}